Fixed favourite bug; fixed cart add bug

+++ app/src/main/java/com/example/final_project/domain/model/restaurant/GetRestaurantDetails.kt
@@ -2,7 +2,7 @@ package com.example.final_project.domain.model.restaurant
 
 data class GetRestaurantDetails(
     val restaurantId: Int,
-    val deliveryFee: Double?,
+    val deliveryFee: Int?,
     val description: String,
     val image: String,
     val latitude: Double,
